feat: 다시 시작하기 버튼을 누르면 초기화 되서 다시 구매를 시작할 수 있다.
