+++ src/js/eventHandler/answer.js
@@ -1,38 +1,7 @@
 import { ANSWER_INPUT_NAMES } from '../constant/lotto.js';
 import lottoManager from '../model/lotto.js';
-import { getLottoAmount } from '../service/lotto.js';
-import {
-  setBriefMode,
-  setDetailMode,
-  updateLottoCount,
-  updateLottoTicketView,
-} from '../view/main.js';
+import { focusPaymentInput, resetMainView } from '../view/main.js';
 import { closeModal, openModal, updateResultView } from '../view/resultModal.js';
-
-const buyLotto = (price) => {
-  const lottoAmount = getLottoAmount(price);
-
-  lottoManager.issue(lottoAmount);
-
-  updateLottoCount(lottoManager.lottos.length);
-  updateLottoTicketView(lottoManager.lottos);
-};
-
-export const handlePayment = (event) => {
-  event.preventDefault();
-
-  const price = event.target.elements['price'].value;
-
-  buyLotto(price);
-};
-
-export const handleLottoDetailToggle = ({ target }) => {
-  if (target.checked) {
-    setDetailMode();
-  } else {
-    setBriefMode();
-  }
-};
 
 const isAnswerValid = (answer) => {
   return Object.values(answer).every((inputNumber) => inputNumber || inputNumber >= 1);
@@ -62,9 +31,8 @@ export const handleAnswer = (event) => {
 
 const reset = () => {
   lottoManager.resetAll();
-  updateLottoCount();
-  updateLottoTicketView();
-  updateResultView();
+  resetMainView();
+  focusPaymentInput();
 };
 
 export const handleModalClose = () => {
